+++ tracing-core/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.12 (UNRELEASED)
+
+## Added
+
+- `LevelFilter` type and `LevelFilter::current()` for returning the highest level
+  that any subscriber will enable (#853)
+- `Subscriber::max_level_hint` optional trait method, for setting the value
+  returned by `LevelFilter::current()` (#853)
+
 # 0.1.11 (June 8, 2020)
 
 ## Changed